LayoutTests/media/video-error-abort.html

<video src=content/test.mp4></video>
<script src=video-test.js></script>
<script>
test("!video.error");
var didLoad = false;
waitForEvent("begin", function () { if (didLoad) return; didLoad = true; video.load(); });
waitForEventAndTest("abort", "video.error.code == MediaError.MEDIA_ERR_ABORTED");
waitForEventTestAndEnd("load", "!video.error");
</script>
